+++ website/sidebars.js
@@ -136,6 +136,7 @@ module.exports = {
             "providers/provider",
             "providers/state_notifier_provider",
             "providers/future_provider",
+            "providers/state_provider",
           ],
         },
         {