+++ packages/riverpod/CHANGELOG.md
@@ -1,3 +1,7 @@
+[Unreleased bugfix]
+
+- Fixed a cast error when overriding a provider with a more specific provider type (#1100)
+
 # 2.0.0-dev.0
 
 - **Breaking** After a provider has emitted an `AsyncValue.data` or `AsyncValue.error`,
@@ -7,6 +11,7 @@
 
   This allows the UI to keep showing the previous data/error when a provider
   is being refreshed.
+
 - Adding `isLoading`, `isError`, `isData` and `asError` to `AsyncValue`.
   Those getters allow interacting with `AsyncValue` without having to rely on
   pattern matching.